pyproject.toml
[project]
name = "lakefront"
version = "0.1.11"
description = "Lakehouse Observability Platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.0.0,<2.0.0",
    "pandas>=2.3.3",
    "pyarrow>=23.0.1",
    "pydantic-settings[toml]>=2.13.1",
    "tomli-w>=1.2.0",
    "textual<=9.0.0",
    "typer>=0.24,<1.0.0",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "deptry>=0.24.0",
    "ipython>=8.38.0",
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]

[project.scripts]
lakefront = "lakefront.cli:app"
lf = "lakefront.cli:app"

[build-system]
requires = ["uv_build>=0.10.3,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = "src"
namespace = true

[tool.pytest.ini_options]
# pythonpath = "pkg/*/src"  
addopts = "-vv --maxfail=1 --tb=short --env-file=test.env"
markers = [
    "integration: tests requiring external services like Docker",
]

[tool.deptry]
ignore = [
    "DEP003"
]

[tool.ruff.lint]
select = ["I", "TID"]

[tool.ruff.lint.isort]
force-sort-within-sections = false
known-first-party = ["lakefront"]
combine-as-imports = true

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"

